While continuing to research and debug segfaults due to null pointer dereferences and use-after-frees, related to `mod_vroot`, for https://github.com/proftpd/proftpd/issues/1384, I determined that these segfaults all occurred during the authentication process.
With this change, `mod_vroot` now registers its custom FS closer to the
(possible) `chroot(2)` system call in timing, and _only_ if authentication
succeeds.  Previously the module would use a PRE_CMD PASS handler to register
the custom FS, then POST_CMD/POST_CMD_ERR PASS handlers to unregister the
FS in case of issues.  This, in turn, means that the custom FS is active (and
used!) _during_ the authentication process, which appeared to be related
to the above memory/reference issues.

So I took a step back, and thought of the purpose of `mod_vroot` (to provide
a custom "virtual chroot" for a session), and changed the timing of the custom
FSIO registration to be closer to the point where it is actually needed/used.